// packages/core/src/types.d.ts
export declare enum AssignmentStatus {
    Pending = "Pendent",
    Yes = "S\u00ED",
    No = "No",
    Mixed = "Mixt"
}
export interface PersonGroup {
    id: string;
    name: string;
    role?: string;
    tel1?: string;
    tel2?: string;
    email?: string;
    web?: string;
    notes?: string;
}
export interface Assignment {
    id: string;
    personGroupId: string;
    eventFrameId: string;
    startDate: string;
    endDate: string;
    status: AssignmentStatus;
    notes?: string;
    dailyStatuses?: {
        [dateYYYYMMDD: string]: AssignmentStatus;
    };
}